requirements.txt
streamlit==1.28.0
ultralytics==8.0.20
opencv-python-headless==4.6.0.66
numpy==1.23.5
Pillow==9.5.0
matplotlib
pandas